requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary 
alembic 
python-dotenv 
pydantic 
pydantic-settings